myworld/requirements.txt
altgraph==0.17.4
asgiref==3.7.2
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
celery==5.4.0
certifi==2024.2.2
cffi==1.17.1
channels==4.2.0
charset-normalizer==3.4.0
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==3.0.0
cmdstanpy==1.2.2
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cramjam==2.8.2
cycler==0.12.1
daphne==4.1.2
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
django-environ==0.12.0
Django==4.2.2
django-preventconcurrentlogins==0.8.2
django-redis==5.4.0
django-session-timeout==0.1.0
django-user-sessions==2.0.0
et-xmlfile==1.1.0
executing==2.0.1
fastjsonschema==2.19.1
fastparquet==2024.2.0
filelock==3.12.0
fiona==1.9.5
fonttools==4.47.2
fsspec==2024.2.0
geopandas==0.14.3
h11==0.14.0
holidays==0.48
idna==3.10
importlib_resources==6.4.0
ipykernel==6.29.2
ipython==8.22.1
jalali_core==1.0.0
jdatetime==5.0.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyterlab_pygments==0.3.0
kiwisolver==1.4.5
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
nbclient==0.9.0
nbconvert==7.16.1
nbformat==5.9.2
nest-asyncio==1.6.0
numpy==1.25.2
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.0
pandas-flavor==0.6.0
pandocfilters==1.5.1
parso==0.8.3
patsy==0.5.6
pefile==2023.2.7
persiantools==3.0.1
pillow==10.2.0
pingouin==0.5.5
platformdirs==3.5.1
prompt-toolkit==3.0.43
prophet==1.1.5
psutil==5.9.8
psycopg2==2.9.9
pure-eval==0.2.2
pycparser==2.22
Pygments==2.17.2
pyinstaller==6.4.0
pyinstaller-hooks-contrib==2024.1
pyparsing==3.1.1
pyproj==3.6.1
pyserial==3.5
PySocks==1.7.1
pyspv1==1.0.2
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3
pyzmq==25.1.2
redis==5.0.1
referencing==0.33.0
requests==2.32.3
rich==13.7.1
rpds-py==0.18.0
savReaderWriter==3.4.2
scikit-learn==1.6.1
scipy==1.12.0
seaborn==0.13.2
selenium==4.26.1
shapely==2.0.2
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
spyder-kernels==2.5.0
sqlparse==0.4.4
stack-data==0.6.3
stanio==0.5.0
statsmodels==0.14.1
tabulate==0.9.0
threadpoolctl==3.5.0
tinycss2==1.2.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.1
trio==0.27.0
trio-websocket==0.11.1
typing==3.7.4.3
typing_extensions==4.12.2
tzdata==2023.3
urllib3==2.2.3
virtualenv==20.23.0
wcwidth==0.2.13
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
wsproto==1.2.0
xarray==2025.1.2
